fix registration datatype

+++ src/main/java/com/javerian/erp/mlm/model/workflow/Registration.java
@@ -1,17 +1,11 @@
 package com.javerian.erp.mlm.model.workflow;
 
-import javax.persistence.CascadeType;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToOne;
-import javax.persistence.PrimaryKeyJoinColumn;
 import javax.persistence.Table;
-
-import org.hibernate.annotations.GenericGenerator;
-import org.hibernate.annotations.Parameter;
 
 @Entity
 @Table(name = "registration")
@@ -21,32 +15,32 @@ public class Registration {
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	private Long id;
 	@Column(name = "registration_amout")
-	private int registration_amout;
+	private Double registration_amout;
 	@Column(name = "remark")
 	private String remark;
+
 	public Long getId() {
 		return id;
 	}
+
 	public void setId(Long id) {
 		this.id = id;
 	}
-	public int getRegistration_amout() {
+
+	public Double getRegistration_amout() {
 		return registration_amout;
 	}
-	public void setRegistration_amout(int registration_amout) {
+
+	public void setRegistration_amout(Double registration_amout) {
 		this.registration_amout = registration_amout;
 	}
+
 	public String getRemark() {
 		return remark;
 	}
+
 	public void setRemark(String remark) {
 		this.remark = remark;
 	}
-	
-	
-	
-	
-	
-	
-	
+
 }
